refactor: Extract single file download logic to custom hook (#8944)
✨ (use-custom-handle-bulk-files-download.ts): add custom hook to handle bulk files download functionality
✨ (use-custom-handle-single-file-download.ts): add custom hook to handle single file download functionality
🔧 (index.tsx): fix import path for use-custom-handle-bulk-files-download hook in filesPage
🔧 (index.tsx): fix import path for use-custom-handle-single-file-download hook in filesContextMenuComponent
