+++ src/frontend/src/customization/hooks/use-custom-handle-single-file-download.ts
@@ -1,0 +1,26 @@
+import { useGetDownloadFileV2 } from "@/controllers/API/queries/file-management";
+import { FileType } from "@/types/file_management";
+
+interface SingleFileDownloadParams {
+  id: string;
+  filename: string;
+  type: string;
+}
+
+export const useCustomHandleSingleFileDownload = (file: FileType) => {
+  const { mutate: downloadFile } = useGetDownloadFileV2({
+    id: file.id,
+    filename: file.name,
+    type: file.path.split(".").pop() || "",
+  });
+
+  const handleSingleDownload = (
+    params?: SingleFileDownloadParams,
+    setSuccessData?: (data: { title: string }) => void,
+    setErrorData?: (data: { title: string; list: string[] }) => void,
+  ) => {
+    downloadFile();
+  };
+
+  return { handleSingleDownload };
+};
