+++ src/frontend/src/modals/fileManagerModal/components/filesContextMenuComponent/index.tsx
@@ -5,9 +5,9 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { useGetDownloadFileV2 } from "@/controllers/API/queries/file-management";
 import { useDeleteFileV2 } from "@/controllers/API/queries/file-management/use-delete-file";
 import { useDuplicateFileV2 } from "@/controllers/API/queries/file-management/use-duplicate-file";
+import { useCustomHandleSingleFileDownload } from "@/customization/hooks/use-custom-handle-single-file-download";
 import ConfirmationModal from "@/modals/confirmationModal";
 import useAlertStore from "@/stores/alertStore";
 import { FileType } from "@/types/file_management";
@@ -29,11 +29,7 @@ export default function FilesContextMenuComponent({
 
   const setSuccessData = useAlertStore((state) => state.setSuccessData);
 
-  const { mutate: downloadFile } = useGetDownloadFileV2({
-    id: file.id,
-    filename: file.name,
-    type: file.path.split(".").pop() || "",
-  });
+  const { handleSingleDownload } = useCustomHandleSingleFileDownload(file);
 
   const { mutate: deleteFile } = useDeleteFileV2({
     id: file.id,
@@ -54,7 +50,7 @@ export default function FilesContextMenuComponent({
         console.log("replace");
         break;
       case "download":
-        downloadFile();
+        handleSingleDownload();
         break;
       case "delete":
         setShowDeleteConfirmation(true);
